+++ tests/DotPulsar.Tests/IntegrationFixture.cs
@@ -62,7 +62,7 @@ public class IntegrationFixture : IAsyncLifetime
 
         _cluster = new Builder()
             .UseContainer()
-            .UseImage("apachepulsar/pulsar:2.10.0")
+            .UseImage("apachepulsar/pulsar:2.10.3")
             .WithEnvironment(environmentVariables)
             .ExposePort(Port)
             .Command("/bin/bash -c", arguments)