+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 
 - Subscription properties can be added when creating a consumer
+- Support ordered asynchronous `Send(...)` through a `SendChannel` for use cases that require very high throughput.
+  The `SendChannel` is accessible through the `IProducer<TMessage>` interface and allows the user to block
+  future `Send(...)` calls with `Complete()` and awaiting outstanding send operations with `await Completion()`.
 
 ## [2.7.0] - 2022-12-08