+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.0] - 2023-01-26
+
+### Added
+
+- When building a Pulsar client you can now specify whether the certificate revocation list is checked during authentication. The default is 'true'
+
 ## [2.8.0] - 2023-01-20
 
 ### Added